requirements.txt
certifi

	chardet

	click

	colorama

	cycler

	docopt

	Flask

	idna

	itsdangerous

	Jinja2

	joblib

	gunicorn

	kiwisolver

	MarkupSafe

	matplotlib

	numpy

	pandas

	Pillow

	pipreqs

	pyparsing

	python-dateutil

	pytz

	requests

	scikit-learn

	scipy

	six

	sklearn

	threadpoolctl

	urllib3

	Werkzeug

	yarg

	mlxtend

	nltk

	flask-login

	

	

	oauthlib

	pyOpenSSL

	keras

	tensorflow-cpu



